requirements.txt
flask==3.1.0
boto3==1.37.9
numpy==2.2.3
pandas==2.2.3
matplotlib==3.10.1
seaborn==0.13.2
scikit-learn==1.6.1
python-dotenv==1.0.1
requests==2.31.0 